+++ PlanformCreator2.py
@@ -5,8 +5,21 @@
 
     Object model overview (a little simplified) 
 
-    tbd.
+    App                                         - root frame 
+        |-- Panel_File                          - file functions
+        |-- Panel_Wing                          - wing data
+                ...                             - ...
 
+        |-- Diagram_Wing                        - wing overview diagram 
+                |-- Item_Wing                   - Pygtgraph Plot item for complete wing
+                |-- Item_Wing_Airfoils          - Pygtgraph Plot item for airfoils of wing 
+                ...                             - ...
+        |-- Diagram_Planform                    - planform diagram
+        ...
+
+        |-- Wing                                - Entry to model 
+                |-- Planform                    - the planform - main object 
+                ...                             - ...
 """
 
 import os
@@ -17,14 +30,15 @@ from pathlib import Path
 from PyQt6.QtCore           import QMargins
 from PyQt6.QtWidgets        import QApplication, QMainWindow, QWidget, QMessageBox, QFileDialog
 from PyQt6.QtWidgets        import QVBoxLayout, QHBoxLayout
-from PyQt6.QtWidgets        import QTabWidget
 from PyQt6.QtGui            import QCloseEvent, QGuiApplication
 
-# let python find the other modules in modules relativ to path of self - ! before python system modules
-# common modules hosted by AirfoilEditor
-sys.path.insert (1,os.path.join(Path(__file__).parent , 'AirfoilEditor_subtree/modules'))
-# local modules
-sys.path.insert (1,os.path.join(Path(__file__).parent , 'modules'))
+# let python find the other modules in modules relativ to path of self -
+# common modules hosted by AirfoilEditor  ! before python system modules and PlanformCreator path 
+sys.path.insert (1,os.path.join(Path(__file__).parent , 'AirfoilEditor_subtree\\modules'))
+
+# local modules - at the end - AirfoilEditor modules do have precedence
+sys.path.append (os.path.join(Path(__file__).parent , 'modules'))
+sys.path.append (os.path.join(Path(__file__).parent , 'modules\\model'))
 
 from wing                   import Wing
 
@@ -44,184 +58,14 @@ import logging
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
+
 #------------------------------------------------
 
-APP_NAME     = "Planform Creator 2"
-APP_VERSION  = "2.0"
+APP_NAME            = "Planform Creator 2"
+APP_VERSION         = "3.0 beta"
+WORKER_MIN_VERSION  = '1.0.5'
 
-TEMPLATE_DIR = "templates"
-
-
-# --------------------- tmp ----------------------------------------------------
-
-
-class Tab_Panel (QTabWidget):
-    """ 
-    Tab Widget as parent for other items 
-    """
-
-    name = "Panel"             # will be title 
-
-    _width  = None
-    _height = None 
-
-
-    def __init__(self,  
-                 parent=None,
-                 width=None, 
-                 height=None, 
-                 **kwargs):
-        super().__init__(parent=parent, **kwargs)
-
-        self._parent = parent
-
-        if width  is not None: self._width = width
-        if height is not None: self._height = height
-
-        # set width and height 
-        Widget._set_width  (self, self._width)
-        Widget._set_height (self, self._height)
-
-        font = self.font() 
-        _font = size.HEADER.value
-        font.setPointSize(_font[0])
-        font.setWeight   (_font[1])  
-        self.setFont(font)
-
-        # see https://doc.qt.io/qt-6/stylesheet-examples.html
-
-        if Widget.light_mode:
-            tab_style = """
-            QTabWidget::pane { /* The tab widget frame */
-                border-top:1px solid #ababab;
-            }
-
-            QTabWidget::tab-bar {
-                left: 400px; /* move to the right by 5px */
-            }
-
-            /* Style the tab using the tab sub-control. Note that
-                it reads QTabBar _not_ QTabWidget */
-            QTabBar::tab {
-                /*background: green; */
-                border: 1px solid #C4C4C3;
-                border-bottom: 0px;                                     /*remove */
-                border-top-left-radius: 3px;
-                border-top-right-radius: 3px;
-                min-width: 40ex;
-                padding: 6px;
-            }
-
-            QTabBar::tab:!selected {
-                margin-top: 2px; /* make non-selected tabs look smaller */
-                background: #e5e5e5
-            }
-                            
-            QTabBar::tab:hover {
-                background: rgba(255, 255, 255, 0.2) /* rgba(255, 20, 147, 0.1); */              
-            }
-
-            QTabBar::tab:selected {
-                background: rgba(255, 255, 255, 0.9) /* background: rgba(255, 20, 147, 0.2); */               
-            }
-
-            QTabBar::tab:selected {
-                /*color: white; */
-                color: #303030;
-                font-weight: 600;
-                border-color: #9B9B9B;
-                border-bottom-color: #C2C7CB; /* same as pane color */
-            }
-            """
- 
-        else: 
-
-            tab_style = """
-            QTabWidget::pane { /* The tab widget frame */
-                border-top:1px solid #505050;
-            }
-
-            QTabWidget::tab-bar {
-                left: 400px; /* move to the right by 5px */
-            }
-
-            /* Style the tab using the tab sub-control. Note that
-                it reads QTabBar _not_ QTabWidget */
-            QTabBar::tab {
-                /*background: green; */
-                border: 1px solid #505050;  
-                border-bottom: 0px;                                     /*remove */
-                border-top-left-radius: 3px;
-                border-top-right-radius: 3px;
-                min-width: 40ex;
-                padding: 6px;
-            }
-
-            QTabBar::tab:!selected {
-                margin-top: 2px; /* make non-selected tabs look smaller */
-                color: #D0D0D0;
-                background: #353535
-            }
-                            
-            QTabBar::tab:hover {
-                background: rgba(255, 255, 255, 0.2) /* rgba(255, 20, 147, 0.1); */             
-            }
-
-            QTabBar::tab:selected {
-                background: rgba(77, 77, 77, 0.9) /* background: rgba(255, 20, 147, 0.2); */                   
-            }
-
-            QTabBar::tab:selected {
-                /*color: white; */
-                color: #E0E0E0;
-                font-weight: 600;
-                border-color: #909090;
-                border-bottom-color: #C2C7CB;   /* same as pane color */
-            }
-            """
-
-
-        self.setStyleSheet (tab_style) 
-
-
-    def __repr__(self) -> str:
-        # overwritten to get a nice print string 
-        return f"<Tab_Panel '{self.name}'>"
-
-
-    def add_tab (self, aWidget : QWidget, name : str = None):
-        """ at an item having 'name' to self"""
-
-        if name is None:
-            name = aWidget.name
-
-        self.addTab (aWidget, name)
-
-
-    def set_tab (self, class_name : str):
-        """ set the current tab to tab with widgets class name"""
-
-        for itab in range (self.count()):
-            if self.widget(itab).__class__.__name__ == class_name:
-                self.setCurrentIndex (itab)
-                return
-
-
-    def set_background_color (self, darker_factor : int | None = None,
-                                    color : QColor | int | None  = None,
-                                    alpha : float | None = None):
-        """ 
-        Set background color of a QWidget either by
-            - darker_factor > 100  
-            - color: QColor or string for new color
-            - alpha: transparency 0..1 
-        """
-        set_background (self, darker_factor=darker_factor, color=color, alpha=alpha)
-
-
-
-
-
+TEMPLATE_DIR        = "templates"
 
 
 #-------------------------------------------------------------------------------
@@ -231,8 +75,6 @@ class Tab_Panel (QTabWidget):
 class App_Main (QMainWindow):
 
     name = APP_NAME  
-
-    WORKER_MIN_VERSION          = '1.0.5'
 
    # Signals 
 
@@ -270,12 +112,9 @@ class App_Main (QMainWindow):
         Win_Util.set_initialWindowSize (self, size_frac= (0.80, 0.70), pos_frac=(0.1, 0.1),
                                         geometry=geometry, maximize=maximize)
 
-
         # Worker for polar generation ready?
 
-        Worker().isReady (__file__, min_version=self.WORKER_MIN_VERSION)
-        # if Worker.ready:
-        #     Worker().clean_workingDir (self.airfoil().pathName)
+        Worker().isReady (__file__, min_version=WORKER_MIN_VERSION)
 
         # if no initial pc2 file, try to get last openend pc2 file
