+++ modules/model/__init__.py
@@ -1,0 +1,1 @@
+# let python find the other modules if they are launched in this directory (for testing)
