+++ modules/pc2_panels.py
@@ -10,6 +10,7 @@ UI panels
 import logging
 
 from PyQt6.QtWidgets        import QDialog
+from PyQt6.QtWidgets        import QTabWidget
 
 from base.widgets           import * 
 from base.panels            import Edit_Panel, Container_Panel
@@ -25,6 +26,177 @@ from AirfoilEditor_subtree.AirfoilEditor import App_Main
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
+
+
+
+#-------------------------------------------------------------------------------
+# Tab panel    
+#-------------------------------------------------------------------------------
+
+
+
+class Tab_Panel (QTabWidget):
+    """ 
+    Tab Widget as parent for other items 
+    """
+
+    name = "Panel"             # will be title 
+
+    _width  = None
+    _height = None 
+
+
+    def __init__(self,  
+                 parent=None,
+                 width=None, 
+                 height=None, 
+                 **kwargs):
+        super().__init__(parent=parent, **kwargs)
+
+        self._parent = parent
+
+        if width  is not None: self._width = width
+        if height is not None: self._height = height
+
+        # set width and height 
+        Widget._set_width  (self, self._width)
+        Widget._set_height (self, self._height)
+
+        font = self.font() 
+        _font = size.HEADER.value
+        font.setPointSize(_font[0])
+        font.setWeight   (_font[1])  
+        self.setFont(font)
+
+        # see https://doc.qt.io/qt-6/stylesheet-examples.html
+
+        if Widget.light_mode:
+            tab_style = """
+            QTabWidget::pane { /* The tab widget frame */
+                border-top:1px solid #ababab;
+            }
+
+            QTabWidget::tab-bar {
+                left: 400px; /* move to the right by 5px */
+            }
+
+            /* Style the tab using the tab sub-control. Note that
+                it reads QTabBar _not_ QTabWidget */
+            QTabBar::tab {
+                /*background: green; */
+                border: 1px solid #C4C4C3;
+                border-bottom: 0px;                                     /*remove */
+                border-top-left-radius: 3px;
+                border-top-right-radius: 3px;
+                min-width: 40ex;
+                padding: 6px;
+            }
+
+            QTabBar::tab:!selected {
+                margin-top: 2px; /* make non-selected tabs look smaller */
+                background: #e5e5e5
+            }
+                            
+            QTabBar::tab:hover {
+                background: rgba(255, 255, 255, 0.2) /* rgba(255, 20, 147, 0.1); */              
+            }
+
+            QTabBar::tab:selected {
+                background: rgba(255, 255, 255, 0.9) /* background: rgba(255, 20, 147, 0.2); */               
+            }
+
+            QTabBar::tab:selected {
+                /*color: white; */
+                color: #303030;
+                font-weight: 600;
+                border-color: #9B9B9B;
+                border-bottom-color: #C2C7CB; /* same as pane color */
+            }
+            """
+ 
+        else: 
+
+            tab_style = """
+            QTabWidget::pane { /* The tab widget frame */
+                border-top:1px solid #505050;
+            }
+
+            QTabWidget::tab-bar {
+                left: 400px; /* move to the right by 5px */
+            }
+
+            /* Style the tab using the tab sub-control. Note that
+                it reads QTabBar _not_ QTabWidget */
+            QTabBar::tab {
+                /*background: green; */
+                border: 1px solid #505050;  
+                border-bottom: 0px;                                     /*remove */
+                border-top-left-radius: 3px;
+                border-top-right-radius: 3px;
+                min-width: 40ex;
+                padding: 6px;
+            }
+
+            QTabBar::tab:!selected {
+                margin-top: 2px; /* make non-selected tabs look smaller */
+                color: #D0D0D0;
+                background: #353535
+            }
+                            
+            QTabBar::tab:hover {
+                background: rgba(255, 255, 255, 0.2) /* rgba(255, 20, 147, 0.1); */             
+            }
+
+            QTabBar::tab:selected {
+                background: rgba(77, 77, 77, 0.9) /* background: rgba(255, 20, 147, 0.2); */                   
+            }
+
+            QTabBar::tab:selected {
+                /*color: white; */
+                color: #E0E0E0;
+                font-weight: 600;
+                border-color: #909090;
+                border-bottom-color: #C2C7CB;   /* same as pane color */
+            }
+            """
+
+
+        self.setStyleSheet (tab_style) 
+
+
+    def __repr__(self) -> str:
+        # overwritten to get a nice print string 
+        return f"<Tab_Panel '{self.name}'>"
+
+
+    def add_tab (self, aWidget : QWidget, name : str = None):
+        """ at an item having 'name' to self"""
+
+        if name is None:
+            name = aWidget.name
+
+        self.addTab (aWidget, name)
+
+
+    def set_tab (self, class_name : str):
+        """ set the current tab to tab with widgets class name"""
+
+        for itab in range (self.count()):
+            if self.widget(itab).__class__.__name__ == class_name:
+                self.setCurrentIndex (itab)
+                return
+
+
+    def set_background_color (self, darker_factor : int | None = None,
+                                    color : QColor | int | None  = None,
+                                    alpha : float | None = None):
+        """ 
+        Set background color of a QWidget either by
+            - darker_factor > 100  
+            - color: QColor or string for new color
+            - alpha: transparency 0..1 
+        """
+        set_background (self, darker_factor=darker_factor, color=color, alpha=alpha)
